requirements.txt
Django>=5.2.0
python-dotenv>=1.0.0
aiogram>=3.1.0
Pillow>=10.0.0
psycopg[binary]>=3.1.8
uvicorn>=0.23.0 
asgiref>=3.7.2 
openpyxl>==3.1.2
yookassa>==3.5.0